+++ src/main/java/courgette/api/testng/TestNGCourgette.java
@@ -49,11 +49,12 @@ public abstract class TestNGCourgette {
         if (courgetteRunner.canRunFeatures()) {
             courgetteRunner.run();
             courgetteRunner.createReport();
-            courgetteRunner.createCourgetteReport();
 
             if (courgetteProperties.isCucumberReportPublisherEnabled()) {
                 courgetteRunner.publishCucumberReport();
             }
+
+            courgetteRunner.createCourgetteReport();
 
             if (courgetteProperties.isExtentReportsPluginEnabled()) {
                 courgetteRunner.createCourgetteExtentReports();