+++ src/main/java/courgette/runtime/CourgetteSystemProperty.java
@@ -9,4 +9,5 @@ final class CourgetteSystemProperty {
     final static String VM_OPTIONS = "courgette.vmoptions";
     final static String REPORT_TITLE = "courgette.reportTitle";
     final static String REPORT_TARGET_DIR = "courgette.reportTargetDir";
+    final static String ENVIRONMENT_INFO = "courgette.environmentInfo";
 }